backend/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.2

# Machine Learning
scikit-learn>=1.3.2
tensorflow>=2.20.0

# Model Persistence
joblib==1.3.2

# Environment & Deployment
python-dotenv==1.0.0
